Split ec openssl wrapper into separate files
