+++ src/Makefile.qt.include
@@ -266,6 +266,7 @@ BITCOIN_QT_CPP = \
   init.cpp \
   dbwrapper.cpp \
   txdb.cpp \
+  ecwrapper.cpp \
   hash.cpp \
   base58.cpp \
   version.cpp \
